kmrl-system/requirements.txt
# KMRL Knowledge Hub - Global Requirements

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database & Cache
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4

# Document Processing
markitdown==0.0.1a0
pytesseract==0.3.10
Pillow==10.1.0
python-magic==0.4.27

# Storage & Vector DB
minio==7.2.0
opensearch-py==2.4.0

# AI & ML
openai==1.3.7
sentence-transformers==2.2.2
numpy==1.24.4
scikit-learn==1.3.2

# Email & IMAP
imaplib2==3.6

# HTTP & API
requests==2.31.0
httpx==0.25.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
